+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+- Dependencies: Updated to sqlalchemy-cratedb 0.43.0
+- Runtime: Removed support for Python 3.9
 
 ## 2026/05/14 v0.0.0
 - Make it work